requirements.txt
tensorflow_hub
sentencepiece
numpy
pandas
scikit-learn
